third_party/WebKit/LayoutTests/http/tests/devtools/runtime/runtime-getProperties.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script>

var object1 = { get foo() { return 1; }, set foo(value) { } };
var object2 = { get foo() { return 1; } };

function test() {
  var obj1, obj2;

  function dumpProperties(next, properties) {
    for (var i = 0; i < properties.length; ++i)
      dumpProperty(properties[i]);
    next();
  }

  TestRunner.runTestSuite([
    async function testSetUp(next) {
      await new Promise(resolve => TestRunner.evaluateInPage('dumpObjects(\'Initial\')', resolve));

      var result = await TestRunner.RuntimeAgent.evaluate('object1');

      obj1 = TestRunner.runtimeModel.createRemoteObject(result);
      var result = await TestRunner.RuntimeAgent.evaluate('object2');

      obj2 = TestRunner.runtimeModel.createRemoteObject(result);
      next();
    },

    function testGetterAndSetter(next) {
      obj1.getOwnProperties(false, dumpProperties.bind(this, next));
    },

    function testGetterOnly(next) {
      obj2.getOwnProperties(false, dumpProperties.bind(this, next));
    }
  ]);

  function convertPropertyValueForTest(propertyObject, fieldName) {
    var value = propertyObject[fieldName];
    if (value)
      propertyObject[fieldName] = {
        type: value.type,
        description: value.description.replace(/^function [gs]et foo/, 'function '),
        objectId: value.objectId
      };
  }

  function dumpProperty(property) {
    if (property.name === '__proto__')
      return;

    convertPropertyValueForTest(property, 'value');
    convertPropertyValueForTest(property, 'getter');
    convertPropertyValueForTest(property, 'setter');
    TestRunner.dump(property, {objectId: 'formatAsTypeName'});
  }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests RemoteObject.getProperties.
</p>

</body>
</html>
