third_party/WebKit/LayoutTests/http/tests/devtools/runtime/runtime-getProperties-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests RemoteObject.getProperties.


Running: testSetUp

Running: testGetterAndSetter
{
    enumerable : true
    getter : {
        description : "get foo() { return 1; }"
        objectId : <string>
        type : "function"
    }
    isOwn : true
    name : "foo"
    setter : {
        description : "set foo(value) { }"
        objectId : <string>
        type : "function"
    }
    synthetic : false
    wasThrown : false
    writable : false
}

Running: testGetterOnly
{
    enumerable : true
    getter : {
        description : "get foo() { return 1; }"
        objectId : <string>
        type : "function"
    }
    isOwn : true
    name : "foo"
    synthetic : false
    wasThrown : false
    writable : false
}
